+++ Homy.Domin/Contract_Repo/IGenric_Repo.cs
@@ -10,10 +10,10 @@ namespace Homy.Domin.Contract_Repo
     {
         public  Task<IEnumerable<T>> GetAllAsync(); 
         public IQueryable<T> GetAll();  // Non-async queryable for EF includes
-        public Task<T> GetByIdAsync(int id);
+        public Task<T> GetByIdAsync(long id);
         public Task<T>AddAsync(T item);
         public  void Update(T item);
-        public Task<T> DeleteAsync(int id);
+        public Task<T> DeleteAsync(long id);
